refactor HotelierDatabase
