+++ app/src/main/java/com/xepicgamerzx/hotelier/user_activities/LoginActivity.java
@@ -12,7 +12,7 @@ import com.google.android.material.textfield.TextInputEditText;
 import com.xepicgamerzx.hotelier.R;
 import com.xepicgamerzx.hotelier.home_page_activities.MainActivity;
 import com.xepicgamerzx.hotelier.storage.dao.UserDao;
-import com.xepicgamerzx.hotelier.storage.hotelier_database.HotelierDatabase;
+import com.xepicgamerzx.hotelier.storage.HotelierDatabase;
 import com.xepicgamerzx.hotelier.storage.user.model.User;
 
 public class LoginActivity extends AppCompatActivity {
